Development levels

+++ src/main/java/com/example/learnpython/user/service/UserService.java
@@ -5,6 +5,6 @@ import org.springframework.http.ResponseEntity;
 
 public interface UserService {
     ResponseEntity<UserInfoResponse> getUserInfo(final String token, final Long userId);
-    void updateUserLevelAndExpById(final long id, final long gainedExp);
+    void updateUserLevelAndExpById(final long id, final int gainedExp);
     void deleteUserByEmail(final String email);
 }
